% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4syahpradana-package.R
\docType{package}
\name{a4syahpradana}
\alias{a4syahpradana-package}
\alias{a4syahpradana}
\title{a4syahpradana: Exploring Common American Names}
\description{
The package contains:
}
\details{
\itemize{
\item \code{\link{common_names}} — cleaned dataset of common American name combinations.
\item \code{\link{top_ranked_names}()} — helper to summarise and rank names.
\item \code{\link{run_names_app}()} — launches the interactive Shiny app.
}
}
\author{
\strong{Maintainer}: Akbar Rakha Syahpradana \email{asya0008@student.monash.edu}

}
